+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Added
+
+- Support for C#'s `packages.*.config` lockfile type
+
 ## 7.1.5 - 2024-11-26
 
 ### Fixed